Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
     int a,b,c,;
-    float d;
+    int d;
     printf("value of a is : ");
     scanf("%d",&a);
     printf("value of b is : ");
@@ -9,6 +9,6 @@ int main(){
     printf("value of c is : ");
     scanf("%d",&c);
     d= (a+b+c)/3;
-    printf("the average of three number is : %f",d);
+    printf("the average of three number is : %d",d);
     return 0;
 }
